// lib/types.ts
export interface ContributorStats {
    username: string;
    avatarUrl: string;
    totalPRs: number;
    mergedPRs: number;
    openPRs: number;
    closedPRs: number;
    isMaintainer: boolean;
}

export interface PullRequest {
    number: number;
    title: string;
    state: string;
    created_at: string;
    merged_at: string | null;
    html_url: string;
    repository_url: string;
    repository_name?: string;
    user: {
        login: string;
        avatar_url: string;
    };
}